requirements.txt
Django==3.1.2
django-admin-sortable2==0.7.7
django-tinymce==3.1.0
Pillow==7.2.0
python-dotenv==0.14.0
